gpu: Use u32 as Index for chunking which slighly improves performance.
fix test and examples

+++ examples/matrix/matmul-gpu/src/lib.rs
@@ -29,7 +29,7 @@ pub fn inner_product_kernel2(a: &[f32], b: &[f32], c: ThreadChunkMatrix2D<'_>, n
                     sum += aa[k] * b[b_idx];
                     b_idx += n;
                 }
-                c[(j, i)] = sum;
+                c[(j as u32, i as u32)] = sum;
             }
             col += dim_x as usize;
         }
@@ -40,7 +40,7 @@ pub fn inner_product_kernel2(a: &[f32], b: &[f32], c: ThreadChunkMatrix2D<'_>, n
 #[cfg(feature = "v1")]
 #[gpu_macros::cuda_kernel]
 pub fn inner_product_kernel(a: &[f32], b: &[f32], c: &mut [f32], n: usize) {
-    let mut c = chunk_mut(c, Map2D::new(n));
+    let mut c = chunk_mut(c, Map2D::new(n as u32));
     let bid_x = block_id::<gpu::DimX>();
     let bid_y = block_id::<gpu::DimY>();
     let tid_x = thread_id::<gpu::DimX>();
@@ -62,7 +62,7 @@ pub fn inner_product_kernel(a: &[f32], b: &[f32], c: &mut [f32], n: usize) {
                     sum += aa[k] * b[b_idx];
                     b_idx += n;
                 }
-                c[(j, i)] = sum;
+                c[(j as u32, i as u32)] = sum;
             }
             col += dim_x as usize;
         }
